+++ docs/setup/changelog.rst
@@ -20,6 +20,7 @@ Upcoming Changes (main branch)
 - Added ``array.copy()`` and ``nd.copy(array)`` functions.
 - Added ``positive`` and ``negative`` functions.
 - Added ``count_nonzero`` functions.
+- Properly bounds check single-value indices.
 
 **Changed**